// GeoPhotoProjec/platforms/android/app/build/intermediates/assets/F0F1/debug/app/models/photoSearchResponse.ts
export class PhotoSearchResponse{
  id: string;
  owner: string;
  secret: string;
  server: string;
  title: string;
  latitude: string;
  longitude: string;
  datetaken: string;
  url_t: string;
  url_m: string;
  url_q: string;
  url_n: string;
  distance: string;

  constructor(){
    this.url_n = " ";
  }
}